+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- Add `fx.Annotated` to allow users to provide named values without creating a new constructor.
+- Add `fx.Annotated` to allow users to provide named values without creating a
+  new constructor.
 
 ## [1.8.0] - 2018-11-06
 ### Added